about: Clarify Freedoom is just raw material.
The previous about page had a good aspect in that it clarified that
Freedoom is just raw artwork and needs a source port to be used.
Rework the text in the first section to try to get some of that back.

Also make some other minor tweaks to wording on this page, and add
a short margin before the section headings.