Information about the in-dev progress of the project and snapshot links.
While people have the best of intentions for reporting bugs, it is a
bit frustrating to hear that a bug was already fixed, but they're
unaware of it, being on the latest official release. Hopefully this
will make it a bit more obvious to how they could check the current
status beforehand.