download: Bump prboom, odamex, add gzdoom.
I like keeping the list minimal (two ports preferred), and to engines
that are relatively popular and simple to use.

GZDoom isn't really simple, but it is popular, compatible with almost
all mods released, and now licensed under the GPLv3 (making it free
software)!

Odamex isn't very popular, few players can be found on multiplayer
servers as a result.

PrBoom-Plus is great but those menus...