Add sentence suggesting that ZIPs can be verified.
It's cumbersome to have either key or hash information on the download
page, so let's just direct people to the assets attached to the release.

Also, a "keys" directory has been added with GPG keys used to sign git
tags. This makes it possible to link to them, if we wish to do so, but
currently the directory is neither mentioned nor linked to.